Packages structure refactored
