+++ src/main/java/com/kodilla/ecommercee/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.kodilla.ecommercee;
+
+public enum OrderStatus {
+    NEW, PAID, DELIVERED;
+}
