+++ src/main/java/com/kodilla/ecommercee/controller/GroupController.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
 import com.kodilla.ecommercee.domain.GroupDto;
 import org.springframework.http.HttpStatus;
