+++ src/main/java/com/kodilla/ecommercee/controller/ProductController.java
@@ -1,4 +1,4 @@
-package com.kodilla.ecommercee;
+package com.kodilla.ecommercee.controller;
 
 import com.kodilla.ecommercee.domain.ProductDto;
 import org.springframework.http.MediaType;
